requirements.txt
asgiref==3.7.2
Django==4.2.4
django-bootstrap5==23.3
django-environ==0.10.0
django-tinymce==3.6.1
Pillow==10.0.0
psycopg2==2.9.7
sqlparse==0.4.4
tzdata==2023.3
gunicorn==21.2.0